% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFF_CAR_2SS_pq_2d.R
\name{CFF_CAR_2_SS_pq}
\alias{CFF_CAR_2_SS_pq}
\title{CFF_CAR_2_SS_pq}
\usage{
CFF_CAR_2_SS_pq(y, d, Phi, A, H, f1, f2, l, pq)
}
\arguments{
\item{y}{Observable data}

\item{d}{Vector of integration orders}

\item{Phi}{Matrix of AR coefficients}

\item{A}{Matrix of factor loadings}

\item{H}{Observation noise variance matrix}

\item{f1}{A vector of length equal the number of fractionally integrated factor
groups. Each entry corresponds to the number of factors for each group, in
descending memory.}

\item{f2}{A scalar that equals the number of short-memory factors}

\item{l}{Scalar order of the AR polynomials for the stationary factors}

\item{pq}{Scalar defining the order of the ARMA approximation for the fractionally
integrated factors. Typically equal to three.}
}
\value{
Returns the state space form of the CFF-CAR model
}
\description{
Function builds the CFF-CAR model as a state space model
}
