% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFF_CAR_LL_pq_2d.R
\name{CFF_CAR_LLval_pq}
\alias{CFF_CAR_LLval_pq}
\title{CFF_CAR_LLval_pq}
\usage{
CFF_CAR_LLval_pq(parm, Y, f1, f2, l, pq)
}
\arguments{
\item{parm}{Vector of parameters. It contains the memory parameters, the
autoregressive parameters for the short memory factors, the factor loadings,
and the variances of the noise terms in the observations equation}

\item{Y}{Observable data}

\item{f1}{A vector of length equal the number of fractionally integrated factor
groups. Each entry corresponds to the number of factors for each group, in
descending memory.}

\item{f2}{A scalar that equals the number of short-memory factors}

\item{l}{Scalar order of the AR polynomials for the stationary factors}

\item{pq}{Scalar defining the order of the ARMA approximation for the fractionally
integrated factors. Typically equal to three.}
}
\value{
Negative log likelihood for the CFF-CAR model
}
\description{
Negative log likelihood for the CFF-CAR model
}
