% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFF_CAR_2SS_pq_2d.R
\name{CFF_CAR_r_obs_pq}
\alias{CFF_CAR_r_obs_pq}
\title{CFF_CAR_r_obs_pq}
\usage{
CFF_CAR_r_obs_pq(gamma, p, f1, f2, pq)
}
\arguments{
\item{gamma}{Contains the parameters. These are the integration orders,
the factor loadings for the long memory factors, and the loadings for the
short memory factors}

\item{f1}{A vector of length equal the number of fractionally integrated factor
groups. Each entry corresponds to the number of factors for each group, in
descending memory.}

\item{f2}{A scalar that equals the number of short-memory factors}

\item{pq}{Scalar defining the order of the ARMA approximation for the fractionally
integrated factors. Typically equal to three.}
}
\value{
Returns the restricted factor loadings
}
\description{
Function places triangular restrictions on loadings to ensure
identification of the factors
}
