% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFF_CAR_2SS_pq_2d.R
\name{CFF_CAR_r_trans_pq}
\alias{CFF_CAR_r_trans_pq}
\title{CFF_CAR_r_trans_pq}
\usage{
CFF_CAR_r_trans_pq(theta, f1, f2, l, pq)
}
\arguments{
\item{theta}{Contains the parameters. These are the the integration orders d,
and the AR coefficients Phi}

\item{f1}{A vector of length equal the number of fractionally integrated factor
groups. Each entry corresponds to the number of factors for each group, in
descending memory.}

\item{f2}{A scalar that equals the number of short-memory factors}

\item{l}{Scalar order of the AR polynomials for the stationary factors}

\item{pq}{Scalar defining the order of the ARMA approximation for the fractionally
integrated factors. Typically equal to three.}
}
\value{
Returns the restricted transition matrix of the fractionally integrated
factor model
}
\description{
Function places restrictions on transition matrix to ensure that
factors are fractionally integrated noise or AR(l) processes
}
