% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFF_CAR_2SS_pq_2d.R
\name{CFF_KS_2_FACTORS_pq}
\alias{CFF_KS_2_FACTORS_pq}
\title{CFF_KS_2_FACTORS_pq}
\usage{
CFF_KS_2_FACTORS_pq(ksmo, d, f1, f2, pq)
}
\arguments{
\item{ksmo}{A matrix holding the smoothed states from the KFAS package}

\item{d}{Vector of integration orders}

\item{f1}{A vector of length equal the number of fractionally integrated factor
groups. Each entry corresponds to the number of factors for each group, in
descending memory.}

\item{f2}{A scalar that equals the number of short-memory factors}

\item{pq}{Scalar defining the order of the ARMA approximation for the fractionally
integrated factors. Typically equal to three.}
}
\value{
Returns the smoothed out factors
}
\description{
Function smoothes out the factors
}
