% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFF_CAR_EM_pq_2d.R
\name{EMstep_CFF_CAR_pq}
\alias{EMstep_CFF_CAR_pq}
\title{EMstep_CFF_CAR_pq}
\usage{
EMstep_CFF_CAR_pq(Y, f1, f2, l, pq, d, Phi, A, H)
}
\arguments{
\item{Y}{Observable data}

\item{f1}{A vector of length equal the number of fractionally integrated factor
groups. Each entry corresponds to the number of factors for each group, in
descending memory.}

\item{f2}{A scalar that equals the number of short-memory factors}

\item{l}{Scalar order of the AR polynomials for the stationary factors}

\item{pq}{Scalar defining the order of the ARMA approximation for the fractionally
integrated factors. Typically equal to three.}

\item{d}{Vector of integration orders}

\item{Phi}{Matrix of AR coefficients}

\item{A}{Matrix of factor loadings}

\item{H}{Observation noise variance matrix}
}
\value{
EM-updated parameters and likelihood
}
\description{
Single Step of the EM algorithm
}
