% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frac_diff.R
\name{frac_diff}
\alias{frac_diff}
\title{frac_diff}
\usage{
frac_diff(x, d)
}
\arguments{
\item{x}{A (vector) time series}

\item{d}{The (scalar) integration order}
}
\value{
The fractionally differenced x
}
\description{
Takes fractional differences of order d
}
\examples{
frac_diff(rnorm(100), -0.75)

}
