% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frac_diff_m.R
\name{frac_diff_m}
\alias{frac_diff_m}
\title{frac_diff_m}
\usage{
frac_diff_m(x, d)
}
\arguments{
\item{x}{A (n x p) matrix, where each column corresponds to a time series}

\item{d}{Either a scalar or a vector of length p defining the amount of differencing}
}
\value{
The fractionally differenced x
}
\description{
Multivariate fractional differencing function
}
\examples{
frac_diff_m(matrix(rnorm(200), 100, 2), c(-0.75, -1.25))

}
