% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_inf.R
\name{ma_inf}
\alias{ma_inf}
\title{ma_inf}
\usage{
ma_inf(d, N)
}
\arguments{
\item{d}{A (scalar) integration order}

\item{N}{An integer giving the length of the expansion}
}
\value{
The power series expansion (Wald coefficients) of the fractional differencing polynomial
}
\description{
Returns the Wald coefficients of the fractional differencing polynomial
}
\examples{
ma_inf(0.75, 10)

}
